Cleaned up globals, added root comment analysis by default, added unicode encoding to output csv
'''''
Globals removed from code body and added to if main: loop
Root comment analysis added and set to default. Root comments represent each user's gut reaction to the submission, and so the sentiment is not coloured by replies or arguments.
Unicode encoding avoids throwing an error when writing unicode to csv